refactor(adapters): relocate tool fetch infra into adapters/driven/tools (hexagonal P4 sweep)
Physical move of the lazy tool-fetch machinery (manifest/fetcher/cosign +
manifest.yaml data) from regis/tools/ into the existing
regis/adapters/driven/tools/ package, merging the Tool/load_manifest
re-export into its __init__. Cosmetic placement only — the layering was
already enforced by import-linter regardless of location.

- resources.files("regis.tools") -> resources.files("regis.adapters.driven.tools")
- pyproject package-data key retargeted so manifest.yaml ships from the new package
- ci-tools-manifest.yml / ci-tools-fetch-smoke.yml path filters + manifest reads retargeted
- compute_tool_hashes.sh comment path updated
