+++ regis/adapters/driven/tools/__init__.py
@@ -1,1 +1,5 @@
-"""Driven adapters that run external tools as subprocesses."""
+"""Driven adapters for external tools: subprocess runner + fetch infra."""
+
+from regis.adapters.driven.tools.manifest import Tool, load_manifest
+
+__all__ = ["Tool", "load_manifest"]
