+++ regis/adapters/driven/tools/manifest.yaml
