--- regis/tools/__init__.py
@@ -1,5 +1,0 @@
-"""Tool fetch infrastructure (manifest + downloader)."""
-
-from regis.tools.manifest import Tool, load_manifest
-
-__all__ = ["Tool", "load_manifest"]
