+++ tests/tools/test_cosign.py
@@ -5,8 +5,12 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from regis.tools.cosign import CosignUnavailable, CosignVerificationFailed, verify_blob
-from regis.tools.manifest import CosignPolicy
+from regis.adapters.driven.tools.cosign import (
+    CosignUnavailable,
+    CosignVerificationFailed,
+    verify_blob,
+)
+from regis.adapters.driven.tools.manifest import CosignPolicy
 
 
 def test_verify_blob_returns_unavailable_when_cosign_not_in_path(monkeypatch, tmp_path):
@@ -19,14 +23,16 @@ def test_verify_blob_returns_unavailable_when_cosign_not_in_path(monkeypatch, tm
 
 
 def test_verify_blob_success(monkeypatch, tmp_path):
-    monkeypatch.setattr("regis.tools.cosign.shutil.which", lambda _n: "/usr/bin/cosign")
+    monkeypatch.setattr(
+        "regis.adapters.driven.tools.cosign.shutil.which", lambda _n: "/usr/bin/cosign"
+    )
     blob = tmp_path / "artifact.bin"
     blob.write_bytes(b"data")
     policy = CosignPolicy(
         issuer="https://token.actions.githubusercontent.com",
         identity_regex="^https://github\\.com/org/.*",
     )
-    with patch("regis.tools.cosign.subprocess.run") as run:
+    with patch("regis.adapters.driven.tools.cosign.subprocess.run") as run:
         run.return_value = MagicMock(returncode=0, stderr="", stdout="")
         verify_blob(blob, "https://r.example.com/bin.tar.gz", policy)
     cmd = run.call_args[0][0]
@@ -44,11 +50,13 @@ def test_verify_blob_success(monkeypatch, tmp_path):
 
 
 def test_verify_blob_failure_uses_stderr(monkeypatch, tmp_path):
-    monkeypatch.setattr("regis.tools.cosign.shutil.which", lambda _n: "/usr/bin/cosign")
+    monkeypatch.setattr(
+        "regis.adapters.driven.tools.cosign.shutil.which", lambda _n: "/usr/bin/cosign"
+    )
     blob = tmp_path / "a.bin"
     blob.write_bytes(b"x")
     policy = CosignPolicy(issuer="i", identity_regex="r")
-    with patch("regis.tools.cosign.subprocess.run") as run:
+    with patch("regis.adapters.driven.tools.cosign.subprocess.run") as run:
         run.return_value = MagicMock(returncode=1, stderr="bad signature", stdout="")
         with pytest.raises(CosignVerificationFailed) as exc:
             verify_blob(blob, "https://r.example.com/bin.tar.gz", policy)
@@ -56,11 +64,13 @@ def test_verify_blob_failure_uses_stderr(monkeypatch, tmp_path):
 
 
 def test_verify_blob_failure_falls_back_to_stdout(monkeypatch, tmp_path):
-    monkeypatch.setattr("regis.tools.cosign.shutil.which", lambda _n: "/usr/bin/cosign")
+    monkeypatch.setattr(
+        "regis.adapters.driven.tools.cosign.shutil.which", lambda _n: "/usr/bin/cosign"
+    )
     blob = tmp_path / "a.bin"
     blob.write_bytes(b"x")
     policy = CosignPolicy(issuer="i", identity_regex="r")
-    with patch("regis.tools.cosign.subprocess.run") as run:
+    with patch("regis.adapters.driven.tools.cosign.subprocess.run") as run:
         run.return_value = MagicMock(returncode=1, stderr="", stdout="stdout error")
         with pytest.raises(CosignVerificationFailed) as exc:
             verify_blob(blob, "https://r.example.com/bin.tar.gz", policy)
